# internal/gen/http/balance/server/encode_decode.go
// Code generated by goa v3.23.4, DO NOT EDIT.
//
// balance HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/oleshko-g/oggophermart/api/design -o internal/

package server

import (
	"context"
	"errors"
	"io"
	"net/http"
	"strconv"
	"strings"

	balance "github.com/oleshko-g/oggophermart/internal/gen/balance"
	service "github.com/oleshko-g/oggophermart/internal/gen/service"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeUploadUserOrderResponse returns an encoder for responses returned by
// the balance UploadUserOrder endpoint.
func EncodeUploadUserOrderResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*balance.UploadUserOrderResult)
		if res.Accepted != nil && *res.Accepted == "yes" {
			w.WriteHeader(http.StatusAccepted)
			return nil
		}
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodeUploadUserOrderRequest returns a decoder for requests sent to the
// balance UploadUserOrder endpoint.
func DecodeUploadUserOrderRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*balance.UploadUserOrderPayload, error) {
	return func(r *http.Request) (*balance.UploadUserOrderPayload, error) {
		var (
			body string
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = goa.MergeErrors(err, goa.ValidatePattern("body", body, "[1-9][0-9]*"))
		if err != nil {
			return nil, err
		}

		var (
			authorization string
		)
		authorization = r.Header.Get("Authorization")
		if authorization == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Authorization", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewUploadUserOrderPayload(body, authorization)
		if strings.Contains(payload.Authorization, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.Authorization, " ", 2)[1]
			payload.Authorization = cred
		}

		return payload, nil
	}
}

// EncodeUploadUserOrderError returns an encoder for errors returned by the
// UploadUserOrder balance endpoint.
func EncodeUploadUserOrderError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "missing_field":
			var res *goa.ServiceError
			errors.As(v, &res)
			w.Header().Set("Goa-Attribute-Name", res.Name)
			w.Header().Set("Goa-Attribute-Id", res.ID)
			w.Header().Set("Goa-Attribute-Message", res.Message)
			{
				val := res.Temporary
				temporarys := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Temporary", temporarys)
			}
			{
				val := res.Timeout
				timeouts := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Timeout", timeouts)
			}
			{
				val := res.Fault
				faults := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Fault", faults)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "Invalid input parameter":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return nil
		case "The order belongs to another user":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusConflict)
			return nil
		case "Invalid order number":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnprocessableEntity)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeListUserOrdersResponse returns an encoder for responses returned by
// the balance ListUserOrders endpoint.
func EncodeListUserOrdersResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*balance.ListUserOrdersResult)
		if res.NoOrders != nil && *res.NoOrders == "yes" {
			w.WriteHeader(http.StatusNoContent)
			return nil
		}
		enc := encoder(ctx, w)
		body := NewListUserOrdersResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeListUserOrdersRequest returns a decoder for requests sent to the
// balance ListUserOrders endpoint.
func DecodeListUserOrdersRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*balance.ListUserOrdersPayload, error) {
	return func(r *http.Request) (*balance.ListUserOrdersPayload, error) {
		var (
			authorization string
			err           error
		)
		authorization = r.Header.Get("Authorization")
		if authorization == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Authorization", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewListUserOrdersPayload(authorization)
		if strings.Contains(payload.Authorization, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.Authorization, " ", 2)[1]
			payload.Authorization = cred
		}

		return payload, nil
	}
}

// EncodeListUserOrdersError returns an encoder for errors returned by the
// ListUserOrders balance endpoint.
func EncodeListUserOrdersError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "missing_field":
			var res *goa.ServiceError
			errors.As(v, &res)
			w.Header().Set("Goa-Attribute-Name", res.Name)
			w.Header().Set("Goa-Attribute-Id", res.ID)
			w.Header().Set("Goa-Attribute-Message", res.Message)
			{
				val := res.Temporary
				temporarys := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Temporary", temporarys)
			}
			{
				val := res.Timeout
				timeouts := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Timeout", timeouts)
			}
			{
				val := res.Fault
				faults := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Fault", faults)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetUserBalanceResponse returns an encoder for responses returned by
// the balance GetUserBalance endpoint.
func EncodeGetUserBalanceResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*balance.GetUserBalanceResult)
		enc := encoder(ctx, w)
		body := NewGetUserBalanceResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetUserBalanceRequest returns a decoder for requests sent to the
// balance GetUserBalance endpoint.
func DecodeGetUserBalanceRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*balance.GetUserBalancePayload, error) {
	return func(r *http.Request) (*balance.GetUserBalancePayload, error) {
		var (
			authorization string
			err           error
		)
		authorization = r.Header.Get("Authorization")
		if authorization == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Authorization", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetUserBalancePayload(authorization)
		if strings.Contains(payload.Authorization, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.Authorization, " ", 2)[1]
			payload.Authorization = cred
		}

		return payload, nil
	}
}

// EncodeGetUserBalanceError returns an encoder for errors returned by the
// GetUserBalance balance endpoint.
func EncodeGetUserBalanceError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "missing_field":
			var res *goa.ServiceError
			errors.As(v, &res)
			w.Header().Set("Goa-Attribute-Name", res.Name)
			w.Header().Set("Goa-Attribute-Id", res.ID)
			w.Header().Set("Goa-Attribute-Message", res.Message)
			{
				val := res.Temporary
				temporarys := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Temporary", temporarys)
			}
			{
				val := res.Timeout
				timeouts := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Timeout", timeouts)
			}
			{
				val := res.Fault
				faults := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Fault", faults)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeWithdrawUserBalanceResponse returns an encoder for responses returned
// by the balance WithdrawUserBalance endpoint.
func EncodeWithdrawUserBalanceResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodeWithdrawUserBalanceRequest returns a decoder for requests sent to the
// balance WithdrawUserBalance endpoint.
func DecodeWithdrawUserBalanceRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*balance.WithdrawUserBalancePayload, error) {
	return func(r *http.Request) (*balance.WithdrawUserBalancePayload, error) {
		var (
			body WithdrawUserBalanceRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if errors.Is(err, io.EOF) {
				return nil, goa.MissingPayloadError()
			}
			var gerr *goa.ServiceError
			if errors.As(err, &gerr) {
				return nil, gerr
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateWithdrawUserBalanceRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			authorization string
		)
		authorization = r.Header.Get("Authorization")
		if authorization == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Authorization", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewWithdrawUserBalancePayload(&body, authorization)
		if strings.Contains(payload.Authorization, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.Authorization, " ", 2)[1]
			payload.Authorization = cred
		}

		return payload, nil
	}
}

// EncodeWithdrawUserBalanceError returns an encoder for errors returned by the
// WithdrawUserBalance balance endpoint.
func EncodeWithdrawUserBalanceError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "missing_field":
			var res *goa.ServiceError
			errors.As(v, &res)
			w.Header().Set("Goa-Attribute-Name", res.Name)
			w.Header().Set("Goa-Attribute-Id", res.ID)
			w.Header().Set("Goa-Attribute-Message", res.Message)
			{
				val := res.Temporary
				temporarys := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Temporary", temporarys)
			}
			{
				val := res.Timeout
				timeouts := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Timeout", timeouts)
			}
			{
				val := res.Fault
				faults := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Fault", faults)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			enc := encoder(ctx, w)
			var body any
			if formatter != nil {
				body = formatter(ctx, res)
			} else {
				body = NewWithdrawUserBalanceUserIsNotAuthenticatedResponseBody(res)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return enc.Encode(body)
		case "Insufficient funds":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusPaymentRequired)
			return nil
		case "Invalid order number":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnprocessableEntity)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeGetWithdrawalsResponse returns an encoder for responses returned by
// the balance GetWithdrawals endpoint.
func EncodeGetWithdrawalsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*balance.GetWithdrawalsResult)
		if res.NoResult != nil && *res.NoResult == "true" {
			w.WriteHeader(http.StatusNoContent)
			return nil
		}
		enc := encoder(ctx, w)
		body := NewGetWithdrawalsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetWithdrawalsRequest returns a decoder for requests sent to the
// balance GetWithdrawals endpoint.
func DecodeGetWithdrawalsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (*balance.GetWithdrawalsPayload, error) {
	return func(r *http.Request) (*balance.GetWithdrawalsPayload, error) {
		var (
			authorization string
			err           error
		)
		authorization = r.Header.Get("Authorization")
		if authorization == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("Authorization", "header"))
		}
		if err != nil {
			return nil, err
		}
		payload := NewGetWithdrawalsPayload(authorization)
		if strings.Contains(payload.Authorization, " ") {
			// Remove authorization scheme prefix (e.g. "Bearer")
			cred := strings.SplitN(payload.Authorization, " ", 2)[1]
			payload.Authorization = cred
		}

		return payload, nil
	}
}

// EncodeGetWithdrawalsError returns an encoder for errors returned by the
// GetWithdrawals balance endpoint.
func EncodeGetWithdrawalsError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(ctx context.Context, err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		var en goa.GoaErrorNamer
		if !errors.As(v, &en) {
			return encodeError(ctx, w, v)
		}
		switch en.GoaErrorName() {
		case "missing_field":
			var res *goa.ServiceError
			errors.As(v, &res)
			w.Header().Set("Goa-Attribute-Name", res.Name)
			w.Header().Set("Goa-Attribute-Id", res.ID)
			w.Header().Set("Goa-Attribute-Message", res.Message)
			{
				val := res.Temporary
				temporarys := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Temporary", temporarys)
			}
			{
				val := res.Timeout
				timeouts := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Timeout", timeouts)
			}
			{
				val := res.Fault
				faults := strconv.FormatBool(val)
				w.Header().Set("Goa-Attribute-Fault", faults)
			}
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "User is not authenticated":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusUnauthorized)
			return nil
		case "Invalid input parameter":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return nil
		case "Internal service error":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusInternalServerError)
			return nil
		case "Not implemented":
			var res *service.GophermartError
			errors.As(v, &res)
			w.Header().Set("goa-error", res.GoaErrorName())
			w.WriteHeader(http.StatusNotImplemented)
			return nil
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalBalanceOrderToOrder builds a value of type *Order from a value of
// type *balance.Order.
func marshalBalanceOrderToOrder(v *balance.Order) *Order {
	res := &Order{
		Number:     v.Number,
		Status:     v.Status,
		Accrual:    v.Accrual,
		UploadedAt: v.UploadedAt,
	}

	return res
}

// marshalBalanceWithdrawalToWithdrawal builds a value of type *Withdrawal from
// a value of type *balance.Withdrawal.
func marshalBalanceWithdrawalToWithdrawal(v *balance.Withdrawal) *Withdrawal {
	res := &Withdrawal{
		Sum:         v.Sum,
		ProcessedAt: v.ProcessedAt,
	}
	if v.Order != nil {
		order := string(*v.Order)
		res.Order = &order
	}

	return res
}