# internal/gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Goa API
    version: "0.2"
servers:
    - url: http://localhost:80
      description: Default server for gophermart
paths:
    /api/user/balance:
        get:
            tags:
                - balance
            summary: GetUserBalance balance
            description: Get user balance
            operationId: balance#GetUserBalance
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetUserBalanceResponseBody'
                            example:
                                current: 500.5
                                withdrawn: 42
                "401":
                    description: 'missing_field: Missing or empty Authorization header'
                    headers:
                        goa-attribute-fault:
                            description: Is the error a server-side fault?
                            schema:
                                type: boolean
                                description: Is the error a server-side fault?
                                example: true
                            example: false
                        goa-attribute-id:
                            description: ID is a unique identifier for this particular occurrence of the problem.
                            schema:
                                type: string
                                description: ID is a unique identifier for this particular occurrence of the problem.
                                example: 123abc
                            example: 123abc
                        goa-attribute-message:
                            description: Message is a human-readable explanation specific to this occurrence of the problem.
                            schema:
                                type: string
                                description: Message is a human-readable explanation specific to this occurrence of the problem.
                                example: parameter 'p' must be an integer
                            example: parameter 'p' must be an integer
                        goa-attribute-name:
                            description: Name is the name of this class of errors.
                            schema:
                                type: string
                                description: Name is the name of this class of errors.
                                example: bad_request
                            example: bad_request
                        goa-attribute-temporary:
                            description: Is the error temporary?
                            schema:
                                type: boolean
                                description: Is the error temporary?
                                example: false
                            example: true
                        goa-attribute-timeout:
                            description: Is the error a timeout?
                            schema:
                                type: boolean
                                description: Is the error a timeout?
                                example: false
                            example: false
                "500":
                    description: 'Internal service error: Internal Server Error response.'
            security:
                - jwt_header_Authorization: []
    /api/user/balance/withdraw:
        post:
            tags:
                - balance
            summary: WithdrawUserBalance balance
            operationId: balance#WithdrawUserBalance
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/WithdrawUserBalanceRequestBody'
                        example:
                            order: "2377225624"
                            sum: 751
            responses:
                "200":
                    description: OK response.
                "401":
                    description: 'missing_field: Missing or empty Authorization header'
                    headers:
                        goa-attribute-fault:
                            description: Is the error a server-side fault?
                            schema:
                                type: boolean
                                description: Is the error a server-side fault?
                                example: false
                            example: true
                        goa-attribute-id:
                            description: ID is a unique identifier for this particular occurrence of the problem.
                            schema:
                                type: string
                                description: ID is a unique identifier for this particular occurrence of the problem.
                                example: 123abc
                            example: 123abc
                        goa-attribute-message:
                            description: Message is a human-readable explanation specific to this occurrence of the problem.
                            schema:
                                type: string
                                description: Message is a human-readable explanation specific to this occurrence of the problem.
                                example: parameter 'p' must be an integer
                            example: parameter 'p' must be an integer
                        goa-attribute-name:
                            description: Name is the name of this class of errors.
                            schema:
                                type: string
                                description: Name is the name of this class of errors.
                                example: bad_request
                            example: bad_request
                        goa-attribute-temporary:
                            description: Is the error temporary?
                            schema:
                                type: boolean
                                description: Is the error temporary?
                                example: true
                            example: true
                        goa-attribute-timeout:
                            description: Is the error a timeout?
                            schema:
                                type: boolean
                                description: Is the error a timeout?
                                example: true
                            example: false
                "402":
                    description: 'Insufficient funds: Insufficient funds'
                "422":
                    description: 'Invalid order number: Invalid order number'
                "500":
                    description: 'Internal service error: Internal Server Error response.'
            security:
                - jwt_header_Authorization: []
...
    /api/user/withdrawals:
        get:
            tags:
                - balance
            summary: GetWithdrawals balance
            operationId: balance#GetWithdrawals
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/Withdrawal'
                                example:
                                    - order: "946"
                                      processed_at: "1985-11-22T12:53:11Z"
                                      sum: 0.40540135764177376
                                    - order: "946"
                                      processed_at: "1985-11-22T12:53:11Z"
                                      sum: 0.40540135764177376
                            example:
                                - order: "946"
                                  processed_at: "1985-11-22T12:53:11Z"
                                  sum: 0.40540135764177376
                                - order: "946"
                                  processed_at: "1985-11-22T12:53:11Z"
                                  sum: 0.40540135764177376
                "204":
                    description: User has no withdrawals
                "400":
                    description: 'Invalid input parameter: Bad Request response.'
                "401":
                    description: 'missing_field: Missing or empty Authorization header'
                    headers:
                        goa-attribute-fault:
                            description: Is the error a server-side fault?
                            schema:
                                type: boolean
                                description: Is the error a server-side fault?
                                example: false
                            example: false
                        goa-attribute-id:
                            description: ID is a unique identifier for this particular occurrence of the problem.
                            schema:
                                type: string
                                description: ID is a unique identifier for this particular occurrence of the problem.
                                example: 123abc
                            example: 123abc
                        goa-attribute-message:
                            description: Message is a human-readable explanation specific to this occurrence of the problem.
                            schema:
                                type: string
                                description: Message is a human-readable explanation specific to this occurrence of the problem.
                                example: parameter 'p' must be an integer
                            example: parameter 'p' must be an integer
                        goa-attribute-name:
                            description: Name is the name of this class of errors.
                            schema:
                                type: string
                                description: Name is the name of this class of errors.
                                example: bad_request
                            example: bad_request
                        goa-attribute-temporary:
                            description: Is the error temporary?
                            schema:
                                type: boolean
                                description: Is the error temporary?
                                example: true
                            example: true
                        goa-attribute-timeout:
                            description: Is the error a timeout?
                            schema:
                                type: boolean
                                description: Is the error a timeout?
                                example: false
                            example: false
                "500":
                    description: 'Internal service error: Internal Server Error response.'
                "501":
                    description: 'Not implemented: Not Implemented response.'
            security:
                - jwt_header_Authorization: []
components:
    schemas:
        GetUserBalanceResponseBody:
            type: object
            properties:
                current:
                    type: number
                    example: 0.44108522507262227
                    format: double
                    minimum: 0
                withdrawn:
                    type: number
                    example: 0.44153550919667955
                    format: double
                    minimum: 0
            example:
                current: 500.5
                withdrawn: 42
            required:
                - current
                - withdrawn
...
        WithdrawUserBalanceRequestBody:
            type: object
            properties:
                order:
                    type: string
                    example: "48"
                    pattern: '[1-9][0-9]*'
                sum:
                    type: number
                    example: 0.9518688666761115
                    format: double
                    exclusiveMinimum: 0
            example:
                order: "2377225624"
                sum: 751
            required:
                - order
                - sum
        Withdrawal:
            type: object
            properties:
                order:
                    type: string
                    description: Unique user order number
                    example: "3"
                    pattern: '[1-9][0-9]*'
                processed_at:
                    type: string
                    example: "1978-10-18T10:26:38Z"
                    format: date-time
                sum:
                    type: number
                    example: 0.8345306465793881
                    format: double
                    exclusiveMinimum: 0
            example:
                order: "946"
                processed_at: "1999-03-29T17:39:11Z"
                sum: 0.6893665961426662
    securitySchemes:
        jwt_header_Authorization:
            type: http
            description: Secures an endpoint by requiring a valid JWT token.
            scheme: bearer
...